chore: refacto user including methods for the current user
